Minor updates. FIPS implemented, partially working.

+++ src/algorithms/montgomery_algorithm.h
@@ -6,6 +6,7 @@
 #define MONTGOMERY_ALGORITHM_H
 #include <tuple>
 
+
 typedef __int128 int128_t;
 
 class MontgomeryAlgorithm {
@@ -18,7 +19,7 @@ public:
     static int128_t monPro(int128_t a_bar, int128_t b_bar, int128_t n, int128_t n_prime, int128_t r);
 
     // Metoda realizująca rozszerzony algorytm Euklidesa
-    static int128_t gcdExtended(int128_t a, int b, int128_t &x, int128_t &y);
+    static int128_t gcdExtended(int128_t a, int128_t b, int128_t &x, int128_t &y);
 
     // Metoda wyliczająca odwrotność modularną
     static int128_t modInverse(int128_t r, int128_t n);
